Move figures to end, prune methods
Plan is to actually submit the supplement as a separate file.
Methods need to be checked for being too terse.
With figures at nice size in text, we are over limit by 1 page.
With figures at small size in text, over by 0.5 page.

I considered moving the paragraph right after the RSA eqs to
the Methods section but decided against because I thought it
was important to include in the main text.